+++ R/tsmarch-package.R
@@ -10,7 +10,7 @@
 #' @importFrom sandwich estfun bwNeweyWest vcovHAC vcovOPG bread
 #' @importFrom nloptr nloptr
 #' @importFrom grDevices hcl.colors heat.colors terrain.colors topo.colors
-#' @importFrom Rsolnp solnp
+#' @importFrom Rsolnp solnp csolnp
 #' @importFrom graphics grid layout lines par axis barplot contour hist image mtext persp points title abline box
 #' @importFrom future.apply future_lapply
 #' @importFrom future future